requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.1
Flask-Migrate==4.0.7
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Werkzeug==3.0.3
email-validator==2.1.1
gunicorn==22.0.0
Jinja2==3.1.4
itsdangerous==2.2.0
click==8.1.7
alembic==1.13.1
whitenoise==6.7.0
